+++ SnakeGame/src/org/academiadecodigo/tailormoons/snake/Main.java
@@ -1,29 +1,26 @@
 package org.academiadecodigo.tailormoons.snake;
 
-import org.academiadecodigo.simplegraphics.keyboard.Keyboard;
-import org.academiadecodigo.simplegraphics.keyboard.KeyboardEvent;
-import org.academiadecodigo.simplegraphics.keyboard.KeyboardEventType;
-import org.academiadecodigo.tailormoons.snake.Keyboard.KeyHandler;
 import org.academiadecodigo.tailormoons.snake.Keyboard.OurKeyboardHandler;
 import org.academiadecodigo.tailormoons.snake.Menu.GameOver;
 import org.academiadecodigo.tailormoons.snake.Menu.StartMenu;
 import org.academiadecodigo.tailormoons.snake.SnakeGame.SnakeGame;
-import org.academiadecodigo.tailormoons.snake.SnakeGame.SnakeGame1P;
 import org.academiadecodigo.tailormoons.snake.SnakeGrid.SnakeGridNormal;
 import org.academiadecodigo.tailormoons.snake.SnakeGrid.SnakeGridObstacles;
 
 public class Main {
-    public static void main(String[] args) throws InterruptedException {
+    public static void main(String[] args) {
 
         OurKeyboardHandler ourKeyboard = new OurKeyboardHandler();
         ourKeyboard.keyMapping();
+
         StartMenu startMenu = new StartMenu();
         ourKeyboard.setKeyHandling(startMenu);
+        startMenu.setHandler(ourKeyboard);
         startMenu.init();
-
         boolean gameOver = false;
 
         while(true) {
+            startMenu.navigationUpdate();
             if(startMenu.getPlayerType() == 1 && startMenu.getGameType()== 1) break;
             else if (startMenu.getPlayerType() == 1 && startMenu.getGameType() == 2) break;
             else if (startMenu.getPlayerType() == 2 && startMenu.getGameType() == 2) break;
@@ -32,28 +29,29 @@ public class Main {
 
         }
             if (startMenu.getPlayerType() == 1 && startMenu.getGameType() == 1) {
-                SnakeGame1P game = new SnakeGame1P(new SnakeGridNormal());
+                System.out.println("check2");
+                SnakeGame game = new SnakeGame(new SnakeGridNormal());
                 game.setPlayerNumber(1);
                 ourKeyboard.setKeyHandling(game);
                 game.init();
                 game.start();
                 if(game.isGameOver()) gameOver = true;
             } else if (startMenu.getPlayerType() == 1 && startMenu.getGameType() == 2) {
-                SnakeGame1P game = new SnakeGame1P(new SnakeGridObstacles());
+                SnakeGame game = new SnakeGame(new SnakeGridObstacles());
                 game.setPlayerNumber(1);
                 ourKeyboard.setKeyHandling(game);
                 game.init();
                 game.start();
                 if(game.isGameOver()) gameOver = true;
             } else if (startMenu.getGameType() == 1 && startMenu.getPlayerType() == 2)  {
-                SnakeGame1P game = new SnakeGame1P((new SnakeGridNormal()));
+                SnakeGame game = new SnakeGame((new SnakeGridNormal()));
                 game.setPlayerNumber(2);
                 ourKeyboard.setKeyHandling(game);
                 game.init();
                 game.start();
                 if(game.isGameOver()) gameOver = true;
             } else if (startMenu.getGameType() == 2 && startMenu.getPlayerType() == 2) {
-                SnakeGame1P game = new SnakeGame1P(new SnakeGridObstacles());
+                SnakeGame game = new SnakeGame(new SnakeGridObstacles());
                 game.setPlayerNumber(2);
                 ourKeyboard.setKeyHandling(game);
                 game.init();
@@ -62,9 +60,9 @@ public class Main {
             }
 
 
-            GameOver start = new GameOver(ourKeyboard);
-            ourKeyboard.setKeyHandling(start);
-            start.show();
+            GameOver over = new GameOver(ourKeyboard);
+            ourKeyboard.setKeyHandling(over);
+            over.show();
     }
 }
